Drop use of go-restruct/restruct
This appears to not have had any recent release and pulls in yaml.v2
which trivy flags as potentially vulnerable. we can leverage
the encoding/binary functionality to read structures similarly to
how ndb backend already does it.

This saves a trivy warning and also allows upgrading of testify
to latest version.