Take endian-ness of the BerkleyDB into account
When reading a BDB from a different architecture (for instance, from
s390x), it can have a different byte ordering. In order to read an
RPM db from any architecture on any architecture, we can inspect the
magic number, and if it's in a particular order, we can detect whether
to use BigEndian instead of LittleEndian.

Signed-off-by: Brad P. Crochet <brad@redhat.com>