cleanup: add TagName() and TypeName helpers to entryInfo
We open code error messages everywhere and we can instead do it in a
generic way by adding the ability to pull printable tag names and types
from the indexEntry.  We use a generator to parse rpmtags.go and create
a map from the integer values to the names.  This also leads to tag
names and type names that are more consistent with the C RPM library.

This commit just adds the infrastructure for generic error messages.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>