cleanup: convert uint16Array callers to ie.ParseUint16Array
Every caller of uint16Array uses the indexEntry's Data field as an
argument.  If we pass the entire indexEntry, we could pull the tag type
checking from the caller making the main loop easier to read.  Rather
than just do a helper, we can add a member function to do it directly.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>